requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
discord.py==1.4.1
idna==2.10
lxml==4.5.2
multidict==4.7.6
pymongo==3.11.0
python-dotenv==0.14.0
requests==2.24.0
soupsieve==2.0.1
urllib3==1.25.10
yarl==1.5.1
